requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Marshmallow==3.20.1
Python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
pytz==2023.3.post1
flask-cors==4.0.0